Put uncapped-plan tests in Swift Testing files
New test files instead of touching the XCTest-based suites, per the
review policy that touched non-UI tests use Swift Testing. Wired into
the cmuxTests target.
