+++ cmuxTests/MachinesPanelModelTests.swift
@@ -1002,29 +1002,3 @@ struct MachinesPanelPaidPlanTests {
         #expect(error.description.contains("Upgrade to cmux Pro"))
     }
 }
-
-@Suite("Cloud machines uncapped paid plan")
-struct MachinesPanelUncappedPlanTests {
-    private func snapshot(activeCount: Int) -> MachinePlanSnapshot? {
-        MachineSnapshotBuilder.planSnapshot(
-            activeCount: activeCount,
-            limits: VMPlanLimits(maxActiveVms: nil, planId: "pro", freeAccessWindowDays: 0)
-        )
-    }
-
-    @Test("A nil ceiling never reaches the limit")
-    func nilCeilingIsNeverAtLimit() {
-        let plan = snapshot(activeCount: 400)
-        #expect(plan?.maxActiveVms == nil)
-        #expect(plan?.isAtLimit == false)
-        #expect(plan?.isSingleMachinePlan == false)
-        #expect(plan?.isPaidPlan == true)
-    }
-
-    @Test("The meter drops the 'of N' when there is no ceiling", arguments: [
-        (7, "7 machines"), (1, "1 machine"), (0, "0 machines"),
-    ])
-    func meterReadsPlainCount(activeCount: Int, expected: String) {
-        #expect(snapshot(activeCount: activeCount)?.countLabel == expected)
-    }
-}
